Rename bin/validate_names -> bin/validate
This also makes the example more general, allowing it to be used
on more inputs.
