+++ fea-rs/src/diagnostic.rs
@@ -2,7 +2,7 @@
 
 use std::ops::Range;
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum Level {
     Error,
     Warning,
